requirements.txt
gradio
faiss-cpu
sentence-transformers
transformers
torch
pickle-mixin
